requirements.txt
pyTelegramBotAPI==3.6.7
google-cloud-ndb==0.2.2